feat: 개인 획득 통합 스코어 API 구현
Jira: S11P31A210-377

+++ backend/src/main/java/com/reviewping/coflo/domain/project/controller/response/ScoreOfWeekResponse.java
@@ -1,3 +1,3 @@
 package com.reviewping.coflo.domain.project.controller.response;
 
-public record ScoreOfWeekResponse(int week, Long score) {}
+public record ScoreOfWeekResponse(Integer week, Integer score) {}
